pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "safemirror"
version = "0.1.0"
description = "Security-focused package mirror with vulnerability scanning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "taylorelley", email = "taylorelley@example.com"}
]
keywords = ["apt", "mirror", "security", "vulnerability", "scanning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Software Distribution",
]
dependencies = [
    "pyyaml>=6.0,<7.0",
    "requests>=2.32.4,<3.0",
    "click>=8.1.0,<9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<9.0",
    "pytest-cov>=4.1.0,<6.0",
    "pytest-mock>=3.11.0,<4.0",
    "pytest-benchmark>=4.0.0,<6.0",
    "black>=24.3.0,<25.0",
    "ruff>=0.4.0",
    "flake8>=6.0.0,<8.0",
    "mypy>=1.5.0,<2.0",
    "bandit[toml]>=1.7.0",
    "types-PyYAML",
    "types-requests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "benchmark: marks tests as benchmarks",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "src/*/__main__.py",
    "src/*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 70
show_missing = true

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | \.eggs
)/
'''

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
]
ignore = [
    "E501",  # line too long (handled by black)
    "S101",  # assert used (OK in tests)
    "S603",  # subprocess call - check for untrusted input
    "S607",  # subprocess partial path
    "B008",  # function call in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106"]  # Allow asserts and hardcoded passwords in tests

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
disallow_untyped_defs = false  # Gradually enable
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.bandit]
exclude_dirs = ["tests", ".venv", "build", "dist"]
skips = ["B101"]  # assert_used - OK in non-test code for invariants
